chore: move exported types from package entry point into a separate file
