+++ packages/actions-global/src/export-types.ts
@@ -1,0 +1,3 @@
+import type { DispatchActionError } from './actions/handle-action-error';
+
+export type TDispatchActionError = DispatchActionError;
