+++ packages/actions-global/src/index.ts
@@ -1,9 +1,5 @@
-import type { DispatchActionError } from './actions/handle-action-error';
-
-export type TDispatchActionError = DispatchActionError;
-
+export { default as version } from './version';
 export * from './actions';
 export * from './hooks';
 export * from './types';
-
-export { default as version } from './version';
+export * from './export-types';
