+++ packages/application-shell-connectors/src/index.ts
@@ -1,21 +1,5 @@
-import type {
-  TProviderProps as ProviderProps,
-  TApplicationContext as ApplicationContext,
-  TNormalizedMenuVisibilities as NormalizedMenuVisibilities,
-  TNormalizedPermissions as NormalizedPermissions,
-  TNormalizedActionRights as NormalizedActionRights,
-  TNormalizedDataFences as NormalizedDataFences,
-} from './components/application-context';
-
-export type TProviderProps<AdditionalEnvironmentProperties extends {}> =
-  ProviderProps<AdditionalEnvironmentProperties>;
-export type TApplicationContext<AdditionalEnvironmentProperties extends {}> =
-  ApplicationContext<AdditionalEnvironmentProperties>;
-
-export type TNormalizedMenuVisibilities = NormalizedMenuVisibilities;
-export type TNormalizedPermissions = NormalizedPermissions;
-export type TNormalizedActionRights = NormalizedActionRights;
-export type TNormalizedDataFences = NormalizedDataFences;
+export { default as version } from './version';
+export * from './export-types';
 
 export {
   Context,
@@ -34,5 +18,3 @@ export {
   ProjectExtensionProviderForImageRegex,
   withProjectExtensionImageRegex,
 } from './components/project-extension-image-regex';
-
-export { default as version } from './version';
