+++ packages/browser-history/src/export-types.ts
@@ -1,0 +1,4 @@
+import type { EnhancedLocation } from 'history-query-enhancer';
+
+// Convenience types
+export type TEnhancedLocation<Q extends {}> = EnhancedLocation<Q>;
