+++ packages/browser-history/src/index.ts
@@ -1,9 +1,4 @@
-// eslint-disable-next-line import/named
-import type { EnhancedLocation } from 'history-query-enhancer';
-
 export { default as version } from './version';
+export * from './export-types';
 
 export { default, createEnhancedHistory } from './enhanced-history';
-
-// Convenience types
-export type TEnhancedLocation<Q extends {}> = EnhancedLocation<Q>;
