+++ packages/i18n/src/export-types.ts
@@ -1,0 +1,3 @@
+import type { Props } from './async-locale-data/async-locale-data';
+
+export type TAsyncLocaleDataProps = Props;
