+++ packages/notifications/src/index.ts
@@ -1,6 +1,7 @@
 export { default as version } from './version';
+export * from './types';
+
 export { default as middleware } from './middleware';
 export { default as reducer } from './reducer';
 export { addNotification, removeNotification } from './action-creators';
 export { ADD_NOTIFICATION, REMOVE_NOTIFICATION } from './action-types';
-export * from './types';
