+++ packages/permissions/src/index.ts
@@ -1,4 +1,5 @@
 export { default as version } from './version';
+
 export { default as RestrictedByPermissions } from './components/restricted-by-permissions';
 export { default as branchOnPermissions } from './components/branch-on-permissions';
 export {
