+++ packages/react-notifications/src/export-types.ts
@@ -1,0 +1,4 @@
+import type { TProps as MapNotificationToComponentProps } from './components/map-notification-to-component';
+
+// Re-export type for convenience
+export type TMapNotificationToComponentProps = MapNotificationToComponentProps;
