+++ packages/react-notifications/src/index.ts
@@ -1,9 +1,6 @@
-import type { TProps as MapNotificationToComponentProps } from './components/map-notification-to-component';
-
-// Re-export type for convenience
-export type TMapNotificationToComponentProps = MapNotificationToComponentProps;
-
 export { default as version } from './version';
+export * from './export-types';
+
 export { default as NotificationProviderForCustomComponent } from './components/map-notification-to-component';
 export { default as Notification } from './components/notification';
 export {
